Enable automatic joining of tracks
Prior to this commit, tracks had to be manually lined by listing
the head/tail (or head/right/left for switches) connection pairs
in the world scene. This was time consuming to program, error prone,
and difficult to maintain.

This change adds an Area2D node to the head and tail of each section
of track. For switches, one of the head nodes of the two track
segments is disabled so they won't connect to each other by
mistake. The Area2D nodes detect overlapping areas of other track
segments and automatically link the tracks together at the start of
the world's rendering. The areas are then disabled.

Because the tracks aren't connected until the physics simulation
begins (because of the use of the areas) a timer is used to add
the train to the tracks after 0.1 seconds. This allows the track
connections to be generated first and the train cars to cover any
needed tracks.